Move common functions to cdli.py.
Both check.py and upload.py need to open and part the catalogue
csv files. Move this and some other common routines into a shared
module.

There's still some duplication in check.py, which needs access
to the underlying DictReader instance to look at column headings
and so on. I couldn't find an elegant way to do that through
the read_catalogue function.